line docs # 129, 128, 106, 107, 108, 109 (#281)
* line docs # 129, 128, 106, 107, 108, 109

* Update developer.md

Fixed links to SDKs

* Update developer.md

Fixed path on 2 SDK links that were broken by other changes.

* Update quickstart.md

Fixed xref link. The sendemailfromtemplate method is a Server API and not an Admin API.

* Update index.md

Fixed XREF link. The method is called "GetTitleNews" no  "ReadTitleNews" and it is a Client API and not an Admin API.