31 files with pickups (#456)
* 31 files with pickups

* Update real-time-analytics-event-queries.md

Made these changes:
- Line 71: This has to have an * inside the quotes. That's the point of the whole comment. So, to show an asterisk and make it bold without messing up the other formatting, we have to use the escape character \
- Line 102: The period (.) will be confused as a wildcard, unless we use quotes to show that it is not part of the file name.
- Line 119: I know it's grammatically correct to have a period here, but, again, it could be confused as part of the event name (and it just looks weird). So, I changed the preceeding colon to a period.

* Update index.md

Line 21:  The "s" should be bolded as part of the segment name.

* Update event-history.md

Lines 49 - 63: (beyond your edit) This section had to be much clearer in how it uses the method and parameter names.

* Update playstream-events.md

Line 53: I know it's grammatically correct to have a period here but, it could be confused as part of the event name. So, I changed the preceeding colon to a period.

* Update ab-testing-with-title-data.md

Line 124: In the other edits in this doc and this PR, the bolding was removed from CloudScript. So, we should be consistent here.

* Update getting-started-with-playFab-and-snowflake.md

Lines 92, 126, 161: In the other edits in this PR, the bolding was removed from JSON. So, we should be consistent here because we are not referring to a label in the UI, just JSON.