18 files with pickups (#464)
* 18 files with pickups

* Update bulk-actions-for-an-entire-player-segment.md

Deleted an extra space in line 122.

* Update using-cloudscript-actions-with-playstream.md

Fixed a double "the" in line 90.

* Update quickstart.md

The two images needed to be switch to match the description sections. Links needed to be added to the Next Steps section that matched our new tutorial structure and naming.

* Update quickstart.md

Fix two links.

* Update quickstart.md

Fixed casing on a couple of tutorial titles in the links.