17 file pickups (#502)
* 17 file pickups

* Update multiplayer-build-lifecycle.md

Line 31-34: Changed formatting of the state changes to match what was changed in 3 and 4. This looks cleaner and just completes the change that you already started.

* Update thunderhead-archived-release-notes.md

Line 27-31: This section is exactly like the section in Line 52 of thunderhead.md. So, both commands can go in an "azurepowershell" code block.

* Update thunderhead.md

Line 54: Changing code block to azurepowershell instead of just powershell.