Daily publish 04-08-2019 (#505)
* Correction in propertyname concatenation (#495)

Corrected property name concatenation where syntax is

`propertyName<_otherPropertyName*>`

* Change references to devrel and support.playfab.com to helloplayfab@microsoft.com (#496)

* Revisions to DiagnosticsQS (#497)

* 9 file pickups (#498)

* Revisions to Datawarehouse (PF Insights) QS. (#499)

* 17 file pickups (#502)

* 17 file pickups

* Update multiplayer-build-lifecycle.md

Line 31-34: Changed formatting of the state changes to match what was changed in 3 and 4. This looks cleaner and just completes the change that you already started.

* Update thunderhead-archived-release-notes.md

Line 27-31: This section is exactly like the section in Line 52 of thunderhead.md. So, both commands can go in an "azurepowershell" code block.

* Update thunderhead.md

Line 54: Changing code block to azurepowershell instead of just powershell.

* 20 files with pickups (#504)

* 20 files with pickups

* Update using-qos-beacons-to-measure-player-latency-to-azure.md

Line 26: UD should be UDP

* Update basics-of-a-playFab-game-server.md

Line 29: Match the case of the command listed in Line 23 and in another tutorial.

* resolved conflicts

* Update locally-debugging-game-servers-and-integration-with-playfab.md

Line 67: Docker should be capitalized here.