9 file pickups (#519)
* 9 file pickups

* Update quickstart.md

Lines 209, 214: Add code tags for the executable statement.
Lines 232, 243: The semi-colon is part of the code line.

* Update quickstart.md

Line 228, 231, 241: Quote and semi-colon are part of the code line.
Line 248, 252: Include rest of line in code statement.

* Update quickstart.md

Line 244: Add tags to code line.
Line 80: The include directories line was going off the page, so I put it in a code block. That lets it wrap naturally and lets users copy it easily.

* Update quickstart.md

Fix problem with code block in line 80.

* Update quickstart.md

Fix formatting in line 81.

* Update quickstart.md

Match the format of the other quickstarts and put code marks around quoted lines in the "Deconstruct the code" section.