Daily publish 04-16-2019 (#522)
* Correction in propertyname concatenation (#495)

Corrected property name concatenation where syntax is

`propertyName<_otherPropertyName*>`

* Change references to devrel and support.playfab.com to helloplayfab@microsoft.com (#496)

* Revisions to DiagnosticsQS (#497)

* 9 file pickups (#498)

* Revisions to Datawarehouse (PF Insights) QS. (#499)

* 17 file pickups (#502)

* 17 file pickups

* Update multiplayer-build-lifecycle.md

Line 31-34: Changed formatting of the state changes to match what was changed in 3 and 4. This looks cleaner and just completes the change that you already started.

* Update thunderhead-archived-release-notes.md

Line 27-31: This section is exactly like the section in Line 52 of thunderhead.md. So, both commands can go in an "azurepowershell" code block.

* Update thunderhead.md

Line 54: Changing code block to azurepowershell instead of just powershell.

* 20 files with pickups (#504)

* 20 files with pickups

* Update using-qos-beacons-to-measure-player-latency-to-azure.md

Line 26: UD should be UDP

* Update basics-of-a-playFab-game-server.md

Line 29: Match the case of the command listed in Line 23 and in another tutorial.

* resolved conflicts

* Update locally-debugging-game-servers-and-integration-with-playfab.md

Line 67: Docker should be capitalized here.

* Tested and a few revisions to Segmentation quickstart. (#506)

* Changes to Player Login doc organization. (#468)

* Migrate more Pri 1 Tutorials

* Add events to the toc in API Reference.

* Update Thunderhead docs with cutomer feedback.

* Add Party to Quickstart Page.

* Changes to Player Login doc organization.

* meta data title change to match title.

* 11 file pickups, conflicts resolved (#508)

* 11 file pickups, conflicts resolved

* Update quickstart.md

Line 204: Missing space after period.

* Update using-shared-group-data.md

Line 18: Missing space after period.

* Update accessing-archived-tournament-results.md

Line 117: Remove "the" before All Players segment.

* Fix issue 10 by removing duplicate paragraph. (#511)

* Migrate more Pri 1 Tutorials

* Add events to the toc in API Reference.

* Update Thunderhead docs with cutomer feedback.

* Add Party to Quickstart Page.

* Remove duplicate paragraph.

* 11 file pickups (#510)

* Update the multiplayer server powershell quickstart tutorial with latest updates. (#514)

* 9 file pickups (#519)

* 9 file pickups

* Update quickstart.md

Lines 209, 214: Add code tags for the executable statement.
Lines 232, 243: The semi-colon is part of the code line.

* Update quickstart.md

Line 228, 231, 241: Quote and semi-colon are part of the code line.
Line 248, 252: Include rest of line in code statement.

* Update quickstart.md

Line 244: Add tags to code line.
Line 80: The include directories line was going off the page, so I put it in a code block. That lets it wrap naturally and lets users copy it easily.

* Update quickstart.md

Fix problem with code block in line 80.

* Update quickstart.md

Fix formatting in line 81.

* Update quickstart.md

Match the format of the other quickstarts and put code marks around quoted lines in the "Deconstruct the code" section.

* Update event-history.md (#521)

We rolled out this feature so taking off the preview tag.