Update xblguidelines.md with more info, XBL feedback (#674)
* Update xblguidelines.md

* Update xblguidelines.md

Made the following edits:
Lines 27 & 31: Headings should be H2 and have a space after them. That's because we didn't have an H2 already and can't go directly from H1 to H3.
Line 25: I changed the URL to HTTPS and confirmed that it goes to the same page. MS policy is to not use HTTP links if at all possible.