Update quickstart.md (#943)
* Update quickstart.md

This code compiled with minimal changes to the code's logic. 

We may want to double check that the namespace is appropriate. 

Also added a note about the new relying on Client to not stray to far from what the tutorial currently says. We may want to completely remove the code relying on ClientModels though, because that note is a little off putting.

* Update quickstart.md

Removed note about changing Client/Server since we no longer need to make a ClientModels/ServerModels distinction. 

renamed "result" to "response" to align our API's syntax.

Removed excess PlayFab namespaces.