Update PlayFab Party reference documentation and release notes for v1.9.1 (#2365)
Changes:
* Update PlayFab Party reference documentation with new region discovery/measurement control enum values and structures introduced in v1.9.1.
* Apply terminology consistency and URL fixes in various PlayFab Party reference documentation, including some mandatory updates required to pass Acrolinx score minimum bar.
* Update Party error codes convenience documentation reference table with all new values since last manual update (including error 0x3030 for v1.9.1), and broadly update existing strings to use contractions to pass Acrolinx minimum bar.
* Add a Party v1.9.1 release note description entry.
- - -
This consumes auto-generated updates to the PlayFab Party reference documentation as originally found in the Party.h header file for the version 1.9.1 release. That primarily consists of new PartyOptions for advanced control over when and how the Party library performs automatic region discovery and connection quality measurements. There were also minor consistency and correctness fixes made in other parts of the reference documentation.

Note: exceptions were made for spurious line-ending changes that the standard tooling produced in some Party(Local)ChatControl methods but that didn't reflect meaningful content change and so are not included here. Also, many documentation-only issues were found in the official v1.9.1 Party.h but that didn't warrant a new build of the binary, so this contains content generated from a header that doesn't exactly match the v1.9.1 tag. Some of the changes for that have already been committed to the source code repo. However there were also many Acrolinx fixes mandated here in Party.h and BumblelionErrors.h strings that have not yet.

A new release note for this Party version has also been created.